+++ openpyxl/formatting/tests/test_formatting.py
@@ -4,25 +4,162 @@
 from io import BytesIO
 
 # compatibility imports
-from openpyxl import Workbook
-from openpyxl.formatting import ConditionalFormatting
-from openpyxl.formatting.rules import ColorScaleRule, CellIsRule, FormulaRule
+from openpyxl.compat import iteritems, OrderedDict
 
 # package imports
+from openpyxl import Workbook
+from openpyxl.formatting.rules import ColorScaleRule, CellIsRule, FormulaRule
 from openpyxl.reader.excel import load_workbook
 from openpyxl.reader.style import read_style_table
 from openpyxl.xml.constants import ARC_STYLE
-from openpyxl.xml.functions import XMLGenerator, tostring
+from openpyxl.xml.functions import tostring
 from openpyxl.writer.worksheet import write_conditional_formatting
 from openpyxl.writer.styles import StyleWriter
 from openpyxl.styles import Color, PatternFill, Font, Border, Side
 from openpyxl.styles import borders, fills, colors
+from openpyxl.formatting import ConditionalFormatting
 
 # test imports
 import pytest
 from zipfile import ZIP_DEFLATED, ZipFile
 from openpyxl.tests.helper import compare_xml
 from openpyxl.collections import IndexedList
+
+
+@pytest.fixture
+def rules():
+    return OrderedDict(
+        [
+            ('H1:H10', [{'priority': 23, }]),
+            ('Q1:Q10', [{'priority': 14, }]),
+            ('G1:G10', [{'priority': 24, }]),
+            ('F1:F10', [{'priority': 25, }]),
+            ('O1:O10', [{'priority': 16, }]),
+            ('K1:K10', []),
+            ('T1:T10', [{'priority': 11, }]),
+            ('X1:X10', [{'priority': 7, }]),
+            ('R1:R10', [{'priority': 13}]),
+            ('C1:C10', [{'priority': 28, }]),
+            ('J1:J10', [{'priority': 21, }]),
+            ('E1:E10', [{'priority': 26, }]),
+            ('I1:I10', [{'priority': 22, }]),
+            ('Z1:Z10', [{'priority': 5}]),
+            ('V1:V10', [{'priority': 9}]),
+            ('AC1:AC10', [{'priority': 2, }]),
+            ('L1:L10', []),
+            ('N1:N10', [{'priority': 17, }]),
+            ('AA1:AA10', [{'priority': 4, }]),
+            ('M1:M10', []),
+            ('Y1:Y10', [{'priority': 6, }]),
+            ('B1:B10', [{'priority': 29, }]),
+            ('P1:P10', [{'priority': 15, }]),
+            ('W1:W10', [{'priority': 8, }]),
+            ('AB1:AB10', [{'priority': 3, }]),
+            ('A1:A1048576', [{'priority': 29, }]),
+            ('S1:S10', [{'priority': 12, }]),
+            ('D1:D10', [{'priority': 27, }])
+        ]
+    )
+
+
+def test_unpack_rules(rules):
+    from .. import unpack_rules
+    assert list(unpack_rules(rules)) == [
+        ('H1:H10', 0, 23),
+        ('Q1:Q10', 0, 14),
+        ('G1:G10', 0, 24),
+        ('F1:F10', 0, 25),
+        ('O1:O10', 0, 16),
+        ('T1:T10', 0, 11),
+        ('X1:X10', 0, 7),
+        ('R1:R10', 0, 13),
+        ('C1:C10', 0, 28),
+        ('J1:J10', 0, 21),
+        ('E1:E10', 0, 26),
+        ('I1:I10', 0, 22),
+        ('Z1:Z10', 0, 5),
+        ('V1:V10', 0, 9),
+        ('AC1:AC10', 0, 2),
+        ('N1:N10', 0, 17),
+        ('AA1:AA10', 0, 4),
+        ('Y1:Y10', 0, 6),
+        ('B1:B10', 0, 29),
+        ('P1:P10', 0, 15),
+        ('W1:W10', 0, 8),
+        ('AB1:AB10', 0, 3),
+        ('A1:A1048576',0 ,29),
+        ('S1:S10', 0, 12),
+        ('D1:D10', 0, 27),
+    ]
+
+
+def test_update(rules):
+    from .. import unpack_rules, ConditionalFormatting
+    cf = ConditionalFormatting()
+    cf.update(rules)
+    assert cf.max_priority == 25
+    assert list(unpack_rules(cf.cf_rules)) == [
+        ('H1:H10', 0, 18),
+        ('Q1:Q10', 0, 12),
+        ('G1:G10', 0, 19),
+        ('F1:F10', 0, 20),
+        ('O1:O10', 0, 14),
+        ('T1:T10', 0, 9),
+        ('X1:X10', 0, 6),
+        ('R1:R10', 0, 11),
+        ('C1:C10', 0, 23),
+        ('J1:J10', 0, 16),
+        ('E1:E10', 0, 21),
+        ('I1:I10', 0, 17),
+        ('Z1:Z10', 0, 4),
+        ('V1:V10', 0, 8),
+        ('AC1:AC10', 0, 1),
+        ('N1:N10', 0, 15),
+        ('AA1:AA10', 0, 3),
+        ('Y1:Y10', 0, 5),
+        ('B1:B10', 0, 24),
+        ('P1:P10', 0, 13),
+        ('W1:W10', 0, 7),
+        ('AB1:AB10', 0, 2),
+        ('A1:A1048576', 0, 25),
+        ('S1:S10', 0, 10),
+        ('D1:D10', 0, 22),
+    ]
+
+
+def test_fix_priorities(rules):
+    from .. import unpack_rules, ConditionalFormatting
+    cf = ConditionalFormatting()
+    cf.cf_rules = rules
+    cf._fix_priorities()
+    assert cf.max_priority == 25
+    assert list(unpack_rules(cf.cf_rules)) == [
+        ('H1:H10', 0, 18),
+        ('Q1:Q10', 0, 12),
+        ('G1:G10', 0, 19),
+        ('F1:F10', 0, 20),
+        ('O1:O10', 0, 14),
+        ('T1:T10', 0, 9),
+        ('X1:X10', 0, 6),
+        ('R1:R10', 0, 11),
+        ('C1:C10', 0, 23),
+        ('J1:J10', 0, 16),
+        ('E1:E10', 0, 21),
+        ('I1:I10', 0, 17),
+        ('Z1:Z10', 0, 4),
+        ('V1:V10', 0, 8),
+        ('AC1:AC10', 0, 1),
+        ('N1:N10', 0, 15),
+        ('AA1:AA10', 0, 3),
+        ('Y1:Y10', 0, 5),
+        ('B1:B10', 0, 24),
+        ('P1:P10', 0, 13),
+        ('W1:W10', 0, 7),
+        ('AB1:AB10', 0, 2),
+        ('A1:A1048576', 0, 25),
+        ('S1:S10', 0, 10),
+        ('D1:D10', 0, 22),
+    ]
 
 
 class TestRule:
@@ -109,7 +246,7 @@ class TestConditionalFormatting(object):
         worksheet = WS()
         worksheet.conditional_formatting.add('C1:C10', {'type': 'expression', 'formula': ['ISBLANK(C1)'],
                                                         'stopIfTrue': '1', 'dxf': {}})
-        worksheet.conditional_formatting.setDxfStyles(self.workbook)
+        worksheet.conditional_formatting._save_styles(self.workbook)
         cfs = write_conditional_formatting(worksheet)
         xml = b""
         for cf in cfs:
@@ -137,7 +274,7 @@ class TestConditionalFormatting(object):
                                         color=Color(colors.BLACK)))
         cf.add('C1:C10', FormulaRule(formula=['ISBLANK(C1)'], font=font, border=border, fill=fill))
         cf.add('D1:D10', FormulaRule(formula=['ISBLANK(D1)'], fill=fill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         assert len(self.workbook.style_properties['dxf_list']) == 2
         assert self.workbook.style_properties['dxf_list'][0] == {'font': font, 'border': border, 'fill': fill}
         assert self.workbook.style_properties['dxf_list'][1] == {'fill': fill}
@@ -183,7 +320,7 @@ class TestConditionalFormatting(object):
         whiteFont = Font(color=Color("FFFFFFFF"))
         worksheet.conditional_formatting.add('A1:A3', CellIsRule(operator='equal', formula=['"Fail"'], stopIfTrue=False,
                                                                  font=whiteFont, fill=redFill))
-        worksheet.conditional_formatting.setDxfStyles(self.workbook)
+        worksheet.conditional_formatting._save_styles(self.workbook)
 
         # First, verify conditional formatting xml
         cfs = write_conditional_formatting(worksheet)
@@ -282,7 +419,7 @@ class TestCellIsRule(object):
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='greaterThan', formula=['U$7'], stopIfTrue=True, fill=redFill))
         cf.add('V10:V18', CellIsRule(operator='>', formula=['V$7'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -308,7 +445,7 @@ class TestCellIsRule(object):
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='greaterThanOrEqual', formula=['U$7'], stopIfTrue=True, fill=redFill))
         cf.add('V10:V18', CellIsRule(operator='>=', formula=['V$7'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -334,7 +471,7 @@ class TestCellIsRule(object):
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='lessThan', formula=['U$7'], stopIfTrue=True, fill=redFill))
         cf.add('V10:V18', CellIsRule(operator='<', formula=['V$7'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -360,7 +497,7 @@ class TestCellIsRule(object):
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='lessThanOrEqual', formula=['U$7'], stopIfTrue=True, fill=redFill))
         cf.add('V10:V18', CellIsRule(operator='<=', formula=['V$7'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -387,7 +524,7 @@ class TestCellIsRule(object):
         cf.add('U10:U18', CellIsRule(operator='equal', formula=['U$7'], stopIfTrue=True, fill=redFill))
         cf.add('V10:V18', CellIsRule(operator='=', formula=['V$7'], stopIfTrue=True, fill=redFill))
         cf.add('W10:W18', CellIsRule(operator='==', formula=['W$7'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -421,7 +558,7 @@ class TestCellIsRule(object):
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='notEqual', formula=['U$7'], stopIfTrue=True, fill=redFill))
         cf.add('V10:V18', CellIsRule(operator='!=', formula=['V$7'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -446,7 +583,7 @@ class TestCellIsRule(object):
                        end_color=Color('FFEE1111'),
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='between', formula=['U$7', 'U$8'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -464,7 +601,7 @@ class TestCellIsRule(object):
                        end_color=Color('FFEE1111'),
                        patternType=fills.FILL_SOLID)
         cf.add('U10:U18', CellIsRule(operator='notBetween', formula=['U$7', 'U$8'], stopIfTrue=True, fill=redFill))
-        cf.setDxfStyles(self.workbook)
+        cf._save_styles(self.workbook)
         rules = cf.cf_rules
         assert 'U10:U18' in rules
         assert len(cf.cf_rules['U10:U18']) == 1
@@ -490,7 +627,7 @@ class TestFormulaRule(object):
             conditional_formatting = ConditionalFormatting()
         worksheet = WS()
         worksheet.conditional_formatting.add('C1:C10', FormulaRule(formula=['ISBLANK(C1)'], stopIfTrue=True))
-        worksheet.conditional_formatting.setDxfStyles(self.workbook)
+        worksheet.conditional_formatting._save_styles(self.workbook)
 
         cfs = write_conditional_formatting(worksheet)
         xml = b""