+++ openpyxl/writer/tests/test_dump.py
@@ -1,14 +1,18 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2014 openpyxl
 
-import pytest
+from datetime import datetime
+from tempfile import NamedTemporaryFile
+import os
 
+import pytest
 from openpyxl.tests.helper import compare_xml
 
 from openpyxl.collections import IndexedList
-from openpyxl.workbook import CALENDAR_WINDOWS_1900
+from openpyxl.date_time import CALENDAR_WINDOWS_1900
 from openpyxl.styles import Style
-
+from openpyxl.styles.fonts import Font
+from openpyxl.comments.comments import Comment
 
 class DummyLocalData:
 
@@ -30,9 +34,10 @@ class DummyWorkbook:
 
 
 @pytest.fixture
-def DumpWorksheet():
+def DumpWorksheet(request):
     from .. dump_worksheet import DumpWorksheet
-    return DumpWorksheet(DummyWorkbook(), "TestWorkSheet")
+    ws = DumpWorksheet(DummyWorkbook(), "TestWorkSheet")
+    return ws
 
 
 def test_ctor(DumpWorksheet):
@@ -151,7 +156,6 @@ def test_close_content(DumpWorksheet):
     expected = "</sheetData></worksheet>"
     content == expected
 
-
 def test_close(DumpWorksheet):
     ws = DumpWorksheet
     ws.write_header()
@@ -196,3 +200,141 @@ def test_cannot_save_twice(DumpWorksheet):
     ws._cleanup()
     with pytest.raises(WorkbookAlreadySaved):
         ws.get_temporary_file(fname)
+
+
+#Integration tests
+
+from openpyxl import Workbook, load_workbook
+from openpyxl.cell import get_column_letter
+
+
+@pytest.fixture
+def temp_file(tmpdir, request):
+
+    test_file = NamedTemporaryFile(mode='w', prefix='openpyxl.',
+                                   suffix='.xlsx', delete=False, dir=tmpdir.dirname)
+    test_file.close()
+    return test_file.name
+
+
+def test_dump_sheet_title(temp_file):
+    wb = Workbook(optimized_write=True)
+    ws = wb.create_sheet(title='Test1')
+    wb.save(temp_file)
+    wb2 = load_workbook(temp_file)
+    ws = wb2.get_sheet_by_name('Test1')
+    assert 'Test1' == ws.title
+
+
+def test_dump_string_table():
+    wb = Workbook(optimized_write=True)
+    ws = wb.create_sheet()
+    letters = [get_column_letter(x + 1) for x in range(10)]
+
+    for row in range(5):
+        ws.append(['%s%d' % (letter, row + 1) for letter in letters])
+    table = list(wb.shared_strings)
+    assert table == ['A1', 'B1', 'C1', 'D1', 'E1', 'F1', 'G1', 'H1', 'I1', 'J1',
+                     'A2', 'B2', 'C2', 'D2', 'E2', 'F2', 'G2', 'H2', 'I2', 'J2',
+                     'A3', 'B3', 'C3', 'D3', 'E3', 'F3', 'G3', 'H3', 'I3', 'J3',
+                     'A4', 'B4', 'C4', 'D4', 'E4', 'F4', 'G4', 'H4', 'I4', 'J4',
+                     'A5', 'B5', 'C5', 'D5', 'E5', 'F5', 'G5', 'H5', 'I5', 'J5',
+                     ]
+
+
+def test_dump_sheet_with_styles(temp_file):
+    wb = Workbook(optimized_write=True)
+    ws = wb.create_sheet()
+    letters = [get_column_letter(x + 1) for x in range(20)]
+    expected_rows = []
+
+    for row in range(20):
+        expected_rows.append(['%s%d' % (letter, row + 1) for letter in letters])
+
+    for row in range(20):
+        expected_rows.append([(row + 1) for letter in letters])
+
+    for row in range(10):
+        expected_rows.append([datetime(2010, ((x % 12) + 1), row + 1) for x in range(len(letters))])
+
+    for row in range(20):
+        expected_rows.append(['=%s%d' % (letter, row + 1) for letter in letters])
+
+    for row in expected_rows:
+        ws.append(row)
+
+    wb.save(temp_file)
+    wb2 = load_workbook(temp_file)
+    ws = wb2.worksheets[0]
+
+    for ex_row, ws_row in zip(expected_rows[:-20], ws.rows):
+        for ex_cell, ws_cell in zip(ex_row, ws_row):
+            assert ex_cell == ws_cell.value
+
+
+def test_open_too_many_files(temp_file):
+    wb = Workbook(optimized_write=True)
+    for i in range(200): # over 200 worksheets should raise an OSError ('too many open files')
+        wb.create_sheet()
+    wb.save(temp_file)
+
+
+def test_dump_with_font(temp_file):
+    from openpyxl.writer.dump_worksheet import WriteOnlyCell
+
+    wb = Workbook(optimized_write=True)
+    ws = wb.create_sheet()
+    user_style = Style(font=Font(name='Courrier', size=36))
+    cell = WriteOnlyCell(ws, value='hello')
+    cell.style = Style(font=Font(name='Courrier', size=36))
+
+    ws.append([cell, 3.14, None])
+    assert user_style in wb.shared_styles
+    wb.save(temp_file)
+
+    wb2 = load_workbook(temp_file)
+    ws2 = wb2[ws.title]
+    assert ws2['A1'].style == user_style
+
+
+def test_dump_with_comment(temp_file):
+    from openpyxl.writer.dump_worksheet import WriteOnlyCell
+
+    wb = Workbook(optimized_write=True)
+    ws = wb.create_sheet()
+    user_comment = Comment(text='hello world', author='me')
+    cell = WriteOnlyCell(ws, value="hello")
+    cell.comment = user_comment
+
+    ws.append([cell, 3.14, None])
+    assert user_comment in ws._comments
+    wb.save(temp_file)
+
+    wb2 = load_workbook(temp_file)
+    ws2 = wb2[ws.title]
+    assert ws2['A1'].comment.text == 'hello world'
+
+
+@pytest.mark.parametrize("method",
+                         [
+                             '__getitem__',
+                             '__setitem__',
+                             'cell',
+                             'range',
+                             'merge_cells'
+                         ]
+                         )
+def test_illegal_method(method):
+    wb = Workbook(write_only=True)
+    ws = wb.create_sheet()
+    fn = getattr(ws, method)
+    with pytest.raises(NotImplementedError):
+        fn()
+
+def test_save_empty_workbook(temp_file):
+    wb = Workbook(write_only=True)
+    assert len(wb.worksheets) == 0
+    wb.save(temp_file)
+
+    wb = load_workbook(temp_file)
+    assert len(wb.worksheets) == 1